+++ automation/terraform/modules/logging/variables.tf
@@ -1,0 +1,4 @@
+variable "project_name" { type = string }
+variable "action_names" { type = set(string) }
+variable "retention_in_days" { type = number }
+variable "tags" { type = map(string) }